% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPSO.R
\name{assess}
\alias{assess}
\title{Update a particle swarm.}
\usage{
assess(sw, delta, indicator)
}
\arguments{
\item{sw}{the swarm produced by 'initSwarm'. (list)}

\item{delta}{the closeness function that takes a matrix and return a
vector. (num mat -> num vec)}

\item{indicator}{an indicator function that takes a matrix, and returns a
boolean vector that indicates the feasibility of each column of the
matrix. (num mat -> bool vec)}
}
\value{
a list with various properties of the swarm.
}
\description{
Update the properties of an particle swarm.
}
