\name{ordinary-internal}

\alias{projpath}

\alias{shrink_knots}

\alias{\%==nc\%}
\alias{\%innc\%}
\alias{\%w/o\%}
\alias{AUEformat}

\alias{CcovFocal}
\alias{GCV_lamVar_covFix_Wrapper}
\alias{Krig_coef_Wrapper}
\alias{R_GCV_lamVar_covFix}
\alias{getFnEvalCount}
\alias{newCSmooth}
\alias{flushCSmoothTable}

\alias{calcKnotsInfo}
\alias{profileByFullHull}
\alias{calcGridRelProfile}
\alias{indepCalcProfileLRforeachPoint}
\alias{calcBounds1D}
\alias{generateNextpointsfromCI}
\alias{toCanonical}
\alias{bbrhull}
\alias{CKrigcoefs}
\alias{Cpredict}
\alias{Hreprrational}
\alias{OKrig}
\alias{LRTfn}
\alias{LowUpFromKnots}
\alias{LowUpfn}
\alias{addSimplexEq}
\alias{addtoedges}
\alias{as.surface}
\alias{bounds1D}
\alias{calcKnotsInfoWrapper}
\alias{canonizeFromKrig}
\alias{checklogicalbounds}
\alias{condaxialS2fromg}
\alias{constrOptimR}
\alias{constrSymOrRefl}
\alias{constrSymOrReflWrapper}
\alias{deleteCSmooth}
\alias{eps}
\alias{euclidian}
\alias{calcGridFromxy}
\alias{findReplicates}
\alias{findglobalMLE}
\alias{fitinbounds}
\alias{formatName}
\alias{from2Ns2Tocanon}
\alias{fromFONKtoanyspace}
\alias{fromCanonToFONK}
\alias{generateFileName}
\alias{generateInitpts}
\alias{generateNewPoints}
\alias{generatePredictor}
\alias{get_os}
\alias{greedyMAXMINwithFixed}
\alias{gridfn}
\alias{groot}
\alias{handlesPointAtEdge}
\alias{insert.at}
\alias{isPointInCHull}
\alias{isoline}
\alias{knotsFromLRTlist}
\alias{knotsFromProfileEdges}
\alias{.ls.objects}
\alias{locatePointinvT}
\alias{lsos}
\alias{lsOK}
\alias{make.surface.grid}
\alias{makeplot}
\alias{makeplottypes}
\alias{maketicks}
\alias{mantissExp}
\alias{massBarycenter}
\alias{matchVertCons}
\alias{message.redef}
\alias{messageNAMED}
\alias{minimiseGCV}
\alias{myprettyW}
\alias{optimWrapper}
\alias{overmessage}
\alias{parse.grid.list}
\alias{plot.OKrig}
\alias{plot.OKriglistplus}
\alias{plotParPoints}
\alias{plot1DprofFrom2D}
\alias{plotSlice}
\alias{pointsToFONK}
\alias{predict.OKrig}
\alias{predict.OKriglistplus}
\alias{predict_surface}
\alias{prettyNamedUserValues}
\alias{prettyPlotMain}
\alias{prettyUserValues}
\alias{prettynum}
\alias{print.OKrig}
\alias{print.OKriglistplus}
\alias{print.bboptim}
\alias{summary.bboptim}
\alias{summary.OKriglistplus}
\alias{profileBySubHull}
\alias{provideCovFnParams}
\alias{provideDevice}
\alias{providePlotFile}
\alias{providefullhull}
\alias{provideHullwFallBack}
\alias{provideVertices}
\alias{prune_by_dist}
\alias{purefn}
\alias{RAMavail}
\alias{rawProfiles}
\alias{redundant.addVeq}
\alias{resetCHull}
\alias{rExpandedHull}
\alias{rhullByEI}
\alias{rhullByvT}
\alias{safeSurface.OKrig}
\alias{sampleStep}
\alias{savePredictor}
\alias{scdd.addHline}
\alias{setExtrapol}
\alias{setHullPrecMode}
\alias{selectFn}
\alias{selectByLR}
\alias{select_ByHull_ByLogic}
\alias{select_ByLogic}
\alias{set_extrapol_dlr_from_LRTs}
\alias{spaMM.colors}
\alias{stop.redef}
\alias{stripclosestpairs}
\alias{stripclosestpairsWrapper}
\alias{subHullWrapper}
\alias{summary.OKrig}
\alias{surface.OKrig}
\alias{symmetricKnots}
\alias{technicolorTitle}
\alias{temp_capture.output}
\alias{toDgmuFromCanonical}
\alias{tofKpredict.nohull}
\alias{tofullKrigingspace}
\alias{tolatt2Ns2}
\alias{toNMratioFromCanonical}
\alias{tomratioFromCanonical}
\alias{tom1overmuFromCanonical}
\alias{tom2overmuFromCanonical}
\alias{toNactNfounderratioFromCanonical}
\alias{toNfounderNancratioFromCanonical}
\alias{toNfounderratioFromCanonical}
\alias{toNratioFromCanonical}
\alias{toTgmuFromCanonical}
\alias{toUserValues}
\alias{transformx}
\alias{userunit}
\alias{verticesBarycenter}
\alias{volTriangulation}
\alias{volTriangulationWrapper}
\alias{rvolTriangulation}
\alias{subsimplices.volTriangulation}
\alias{rsimplex}
\alias{writeCleanNAMED}
\alias{writeoutput}
\alias{zoomProfile}

\title{Internal ordinary Functions}
\description{
  Internal ordinary functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{internal}
