% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPSO.R
\name{CEPSO}
\alias{CEPSO}
\title{Constrained minimisation with CEPSO}
\usage{
CEPSO(starting, objf, pilot, indicator, neighbour = N/5, iter = 2000,
  N = 100, verbose = FALSE)
}
\arguments{
\item{starting}{a matrix of starting values with each column corresponds to a
particle. (num mat)}

\item{objf}{the objective function that takes a matrix with each column
corresponds to a particle, and return a vector with respective objective
function values. (num mat -> num mat)}

\item{pilot}{a pilot estimate vector. (num vec)}

\item{indicator}{an indicator function that takes a particle matrix and
return a boolean vector, each columnm of which corresponds to a column of
the input matrix. (num mat -> bool vec)}

\item{neighbour}{the number of neighbour (num)}

\item{iter}{iteration number. (num)}

\item{N}{number of total particles. (num)}

\item{verbose}{show the algorithm progression? (bool)}
}
\value{
a list that contains the best particle and its objective function
    value.
}
\description{
Minimise an objective, subject to the constraint specified by an indicator
function.
}
