% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPSO.R
\name{deltaFac}
\alias{deltaFac}
\title{Factory of closeness function.}
\usage{
deltaFac(pilot, objf)
}
\arguments{
\item{pilot}{an pilot estimate. (num vec)}

\item{objf}{the objective function. See CEPSO for more details. (num mat ->
num vec)}
}
\value{
a 'closeness' function that takes in a particle matrix and returns a
    vector of delta values.
}
\description{
The "closeness" function described in the paper. It is defined as the
difference of the objective function values between pilot and current
states.
}
