% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{eval_pols}
\alias{eval_pols}
\title{Evaluate polynomials.}
\usage{
eval_pols(poly_coefs, x)
}
\arguments{
\item{poly_coefs}{a matrix of polynomial coefficients in ascending order,
each column corresponds to a polynomial. (num mat)}

\item{x}{a vector of position where the polynomials are evaluated at. (num vec)}
}
\value{
a matrix, each row and column corresponds to a point in \code{x} and
    a polynomial respectively.
}
\description{
The polynomials are evaluated at given points using the Horner scheme.
}
