% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{get_deriv_nums}
\alias{get_deriv_nums}
\title{Get the numerators of the derivatives of a set of rational functions.}
\usage{
get_deriv_nums(fs, gs, EPS = 1e-06)
}
\arguments{
\item{fs}{a matrix of numerator polynomial in ascending order representing,
each column corresponds to a polynomial. (num mat)}

\item{gs}{a matrix of denominator coefficients in ascending order, each
column corresponds to a polynomial. (num mat)}

\item{EPS}{error tolerance. (num)}
}
\value{
a matrix of polynomial coefficients.
}
\description{
Assuming that the rational function is r(x) = f(x) / g(x), this function
returns f'(x)g(x) - f(x)g'(x).
}
