% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{get_feasibles}
\alias{get_feasibles}
\title{Generate feasible states}
\usage{
get_feasibles(theta, indicator, N = NCOL(theta), dist = rcauchy,
  dist_para = list(scale = 2))
}
\arguments{
\item{theta}{a matrix with each column corresponding to an (unfeasible)
state. (num mat)}

\item{indicator}{an indicator function that takes a matrix and return a
boolean vector, each element of which corresponds to a column of the
input matrix, indicating its feasibility . (num mat -> bool vec)}

\item{N}{the number of feasible states.}

\item{dist}{the distribution of the random noise. This must be a sampler such
as rnorm, rcauchy and rt. (function)}

\item{dist.para}{a list of parameters used in \code{dist}. (list)}
}
\value{
a list consisting of a matrix of transitioned states, a vector of
    objective function values corresponding to the transitioned states, and
    the acceptance proportion (useful for diagnostic purpose).
}
\description{
Generate feasible states by constantly adding random noises to supplied
(non-feasible) states.
}
\details{
This is a rejection sampler that repeatedly adds noises to \code{theta}
feasible states are attained (i.e. when \code{indicator} returns 1). If
\code{N < ncol(theta)}, N states will be randomly sampled from \code{theta},
whereas if \code{N > ncol(theta)}, additional states will be sampled from
\code{theta} to make up the number.

This function will terminate after 10000 of failed trials.
}
