% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyper}
\alias{hyper}
\title{Data from a hyperbolic tangent function.}
\format{A data frame with 30 rows and 2 columns (x and y).
\describe{
  \item{x}{30 equidistant numbers from 0 to 6.}
  \item{y}{responses corresponds to each x.}
}}
\usage{
hyper
}
\description{
A simulated dataset from a hyperbolic tangent function y = 1 + tanh(x - 3) + eps, where 'eps' is a Gaussian noise with 0 mean and 0.3 standard deviation.
}
\keyword{datasets}
