% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyper_out}
\alias{hyper_out}
\title{Data from a hyperbolic tangent function with two outliers.}
\format{A data frame with 30 rows and 2 columns (x and y).
\describe{
  \item{x}{30 equidistant numbers from 0 to 6.}
  \item{y}{responses corresponds to each x.}
}}
\usage{
hyper_out
}
\description{
A simulated dataset from a hyperbolic tangent function y = 1 + tanh(x - 3) + eps, where 'eps' is a Gaussian noise with 0 mean and 0.3 standard deviation. This dataset is corrupted with two outliers (2nd and 29th rows).
}
\keyword{datasets}
