% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPSO.R
\name{initSwarm}
\alias{initSwarm}
\title{Initialise particle swarm.}
\usage{
initSwarm(starting, delta, indicator)
}
\arguments{
\item{starting}{a matrix of starting values. (num mat)}

\item{delta}{the closeness function that takes a matrix and return a
vector. (num mat -> num vec)}

\item{indicator}{an indicator function that takes a matrix, and returns a
boolean vector that indicates the feasibility of each column of the
matrix. (num mat -> bool vec)}
}
\value{
a list with various properties of the swarm.
}
\description{
Initiate a swarm of particles with the given starting values.
}
