% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{is_monotones_fac}
\alias{is_monotones_fac}
\title{Is there any roots in these polynomials? (factory)}
\usage{
is_monotones_fac(dim_f, dim_g, xmin = 0, xmax = Inf,
  increasing = TRUE, EPS = 1e-06)
}
\arguments{
\item{dim_f}{the number of coefficients in the numerator. (num)}

\item{dim_g}{the number of coefficients in the denominator, excluding the
constant term. (num)}

\item{xmin}{the lower bound. (num)}

\item{xmax}{the upper bound. (num)}

\item{increasing}{set to \code{TRUE} when checking for monotonic increasing,
\code{FALSE} for monotonic decreasing.}

\item{EPS}{error tolerance. (num)}
}
\value{
a function that takes in a matrix of rational function coefficients
    (numerator precede denominator, in ascending order), each column
    corresponds to a rational function, and returns a boolean vector
    indicating whether the condition in the description is met.
}
\description{
Produce a function that checks whether a given set of rational functions is
monotone in a given interval.
}
