% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{is_no_roots}
\alias{is_no_roots}
\title{Is there any roots in these polynomials?}
\usage{
is_no_roots(poly_coefs, a, b, EPS = 1e-06)
}
\arguments{
\item{poly_coefs}{a matrix of polynomial coefficients in ascending order,
each column corresponds to a polynomial. (num mat)}

\item{a}{the lower bound. (num)}

\item{b}{the upper bound. (num)}

\item{EPS}{error tolerance. (num)}
}
\value{
a boolean vector indicating whether the condition in the description
    is met.
}
\description{
For a given set of polynomials, find all roots in a given interval and return
\code{TRUE} if there are no roots in it.
}
