% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{is_positives}
\alias{is_positives}
\title{Are these polynomials positive?}
\usage{
is_positives(poly_coefs, a, b, positive = TRUE, EPS = 1e-06)
}
\arguments{
\item{poly_coefs}{a matrix of polynomial coefficients in ascending order,
each column corresponds to a polynomial. (num mat)}

\item{a}{the lower bound. (num)}

\item{b}{the upper bound. (num)}

\item{positive}{if it is \code{FALSE}, check if the polynomial is
negative. (bool)}

\item{EPS}{error tolerance. (num)}
}
\value{
a boolean vector indicating whether the condition in the description
    is met.
}
\description{
Check if a set of polynomial is positive (or negative) inside a given
interval.
}
\details{
This can be done by checking wherether the roots of the polynomials have even
multiplicities. In addition, one of the points in the interval must have the
desired sign.
}
