% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-function.R
\name{rational_rss_fac}
\alias{rational_rss_fac}
\title{Residual sum of squares of rational function models}
\usage{
rational_rss_fac(y, x, dim_f, dim_g)
}
\arguments{
\item{y}{a response vector. (num vec)}

\item{x}{a predictor vector. (num vec)}

\item{dim_f}{the number of coefficients in the numerator. (num)}

\item{dim_g}{the number of coefficients in the denominator, excluding the
constant term. (num)}
}
\value{
a function that takes in a matrix of rational function coefficients
    (numerator precede denominator, in ascending order), each column
    corresponds to a rational function, and returns a vector of redisual sum
    of squares.
}
\description{
Produce a function that calculates the residual sum of squares of a set of
rational function models.
}
\details{
The produced function can be minimised using SMC-SA, multiSA or CEPSO.
}
