% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{rnorm_rw_fac}
\alias{rnorm_rw_fac}
\title{Random walk normal distribution (factory)}
\usage{
rnorm_rw_fac(sd = 1, fraction = 0.97)
}
\arguments{
\item{sd}{the standard deviation of the normal distribution at the first
iteration, that is when \code{k = 1}. (num)}

\item{fraction}{the percentage of standard deviation (relative to the
previous iteration) left after each iteration. (num)}
}
\value{
a function that takes a matrix of current states (each column
    corresponds to a state) (num mat) and the current iteration count (num),
    and produces a matrix of proposed states (num mat).
}
\description{
This is a factory that creates a random walk normal distribution that adds
some noise to the provided values.
}
\details{
This proposal is designed to decrease after each iteration. The default
option is to decrese 3 percentage from the standard deviation at the previous
iteration.
}
