% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA-based.R
\name{SAmove}
\alias{SAmove}
\title{SA-Move (Markov kernel)}
\usage{
SAmove(theta, objf, proposal, temp, k, objv = NULL)
}
\arguments{
\item{theta}{a matrix with each column corresponding to a current state. (num
mat)}

\item{objf}{an objective function that takes a matrix with each column
corresponds to a state, and return a vector with respective objective
function values. (num mat -> num vec)}

\item{proposal}{a random walk proposal that takes a matrix, and produces a
matrix of proposed states, centred on the corresponding columns of the
input matrix. It should also take in the current iteration count. ((num
mat, num) -> num mat)}

\item{temp}{temperature of the current iteration. (num)}

\item{k}{current iteration count. (num)}

\item{objv}{an optional vector specifying the objective function values for
each column in \code{theta}. (num vec)}
}
\value{
a list consisting of a matrix of transitioned states, a vector of
    objective function values corresponding to the transitioned states, and
    the acceptance proportion (useful for diagnostic purpose).
}
\description{
Performs a Markov transition step for each state. This function is intended
for minimisation.
}
