% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_failures.R
\name{trace_failures}
\alias{trace_failures}
\title{Trace the failing objects of a function}
\usage{
trace_failures(fun, args, drop_unchanged_args = FALSE)
}
\arguments{
\item{fun}{The function to trace.}

\item{args}{An optional list of argument names to trace - by default, all arguments
are traced throughout the function.}

\item{drop_unchanged_args}{Whether arguments that stay constant should change,
defaults to **FALSE**.}
}
\value{
A list of all the objects that were changed, including the line
numbers denoting the lines where the changes happened, or a full trace of all
of the changes to arguments throughout the function (if drop_unchanged_args is
set to **TRUE**).
}
\description{
Trace back the manipulations of object inputs to a function
}
