% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPSO.R
\name{updateInertia}
\alias{updateInertia}
\title{Update the inertia "w" for v1.}
\usage{
updateInertia(in_boundary)
}
\arguments{
\item{in_boundary}{a boolean vector specifying whether the particles are
within the boundary. (bool vec)}
}
\description{
Update the inertia "w" for v1.
}
